Setup a json renderer to match the rough logic of the markdown renderer.
